Refactor categories to use fragments
