+++ graphQL/queries/categories.ts
@@ -1,32 +1,41 @@
 import { gql } from 'graphql-request';
+import { IMAGE_FRAGMENT } from '../fragments/image';
+
+/* Fragments */
+import { CATEGORY_DETAILS_FRAGMENT } from '../fragments/categories';
+import { PRODUCT_DETAILS_FRAGMENT } from '../fragments/products';
 
 export const GET_CATEGORIES_LIST = gql`
 	query getCategoriesList($first: Int) {
 		categoriesLists(first: $first) {
-			id
-			name
-			categoryLink
+			...CategoryDetails
+
 			categoryImage {
-				url
+				...ImageDetails
 			}
 		}
 	}
+
+	${CATEGORY_DETAILS_FRAGMENT}
+	${IMAGE_FRAGMENT}
 `;
 
 export const GET_CATEGORY_DETAILS_BY_LINK = gql`
 	query getCategoryDetails($categoryLink: String!) {
 		categoriesList(where: { categoryLink: $categoryLink }) {
-			id
-			name
-			categoryLink
+			...CategoryDetails
+
 			products {
-				id
-				name
-				slug
+				...ProductDetails
+
 				thumbnail {
-					url
+					...ImageDetails
 				}
 			}
 		}
 	}
+
+	${CATEGORY_DETAILS_FRAGMENT}
+	${PRODUCT_DETAILS_FRAGMENT}
+	${IMAGE_FRAGMENT}
 `;
